% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hmcfcop}
\alias{hmcfcop}
\title{HMC inference for factor copula models}
\usage{
hmcfcop(data_, init_, other_)
}
\value{

}
\description{
\code{vifcop} returns NUTS estimations.
}
\examples{
hmcfcop(data, init, other)

\dontrun{
hmcfcop(data, init, other)
}
}
