% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Student_deriv_nu}
\alias{Student_deriv_nu}
\title{Variational inference for factor copula models}
\usage{
Student_deriv_nu(u, v, par, nu)
}
\description{
\code{vifcop} returns variational estimations.
}
\examples{
vifcop(data, init, other)

\dontrun{
vifcop(data, init, other)
}
}
