% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vifcop}
\alias{vifcop}
\title{Variational inference for factor copula models}
\usage{
vifcop(data_, init_, other_)
}
\description{
\code{vifcop} returns variational estimations.
}
\examples{
vifcop(data, init, other)

\dontrun{
vifcop(data, init, other)
}
}
