% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vifcopula-package.R
\docType{package}
\name{vifcopula-package}
\alias{vifcopula-package}
\alias{vifcopula}
\title{Variational Inference of factor copulas}
\description{
Variational Inference of factor copulas
}
\details{
\tabular{ll}{
Package: \tab vifcopula\cr
Type: \tab Package\cr
Version: \tab 0.0.1 \cr
Date: \tab 2017-02-07\cr
License: \tab GPL (>=2)\cr
Depends: \tab R (\eqn{\geq 2.11.0}{>= 2.11.0})\cr
Imports: \tab Rcpp \cr
Suggests: \tab  \cr
LazyLoad: \tab yes }
}
\references{

}
\author{
Hoang Nguyen
}
